Fix #357: Avoid two steps when jumping to decl
The fix consists in using the internal vim machinery to translate a
point/offset to a position in terms of lines and columns. With this
information, we can use the `set_cursor` operation which is not added to
the jump list, unlike the `goto` operation that we were using before.

This commit also clarifies the effect of these operations in the
pertinent docs of every function.